# src/swaggerExpressProject/api/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "NodeJS Mentoring Programm"
  version: "1.0.0"
  title: "NodeJS Mentoring Programm"
  contact:
    email: "bhargava_chittapudi@epam.com"
# host: "localhost:10010"
# basePath: "/api"
...
paths:
...
  /products/{id}:
    x-swagger-router-controller: "products"
    get:
      tags:
      - "products"
      summary: "Get one product"
      description: "Get one product"
      operationId: "getProduct"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of a product"
        required: true
        type: "string"
      responses:
        200:
          description: "search results matching criteria"
          schema:
            $ref: "#/definitions/Product_response"
        400:
          description: "Error"
          schema: 
            type: "object"
    delete:
      tags:
      - "products"
      summary: "Delete product"
      description: "Delete product"
      operationId: "deleteProduct"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of a product"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema: 
            type: "object"
        400:
          description: "Error"
          schema: 
            type: "object"
  /products/{id}/reviews:
    x-swagger-router-controller: "products"
    get:
      tags:
      - "products"
      summary: "Get all product's reviews"
      description: "Get all product's reviews"
      operationId: "getProductReviews"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of a product"
        required: true
        type: "string"
      responses:
        200:
          description: "search results matching criteria"
          schema:
            type: "array"
            items:
              type: "object"
              properties: {}
        400:
          description: "Error"
...
  /cities/random:
    x-swagger-router-controller: "cities"
    get:
      tags:
      - "cities"
      summary: "Get random city"
      description: "Get random city"
      operationId: "getRandomCity"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "OK"
          schema:
            $ref: "#/definitions/City_response"
        400:
          description: "Error"
          schema: 
            type: "object"
  /cities/{id}:
    x-swagger-router-controller: "cities"
    put:
      tags:
      - "cities"
      summary: "Update city info"
      description: "Update city info"
      operationId: "updateCity"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of a city"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "data to update"
        required: false
        schema:
          $ref: "#/definitions/City_body"
      responses:
        200:
          description: "search results matching criteria"
          schema:
            type: "object"
        400:
          description: "Error"
          schema: 
            type: "object"
    delete:
      tags:
      - "cities"
      summary: "Delete city"
      description: "Delete city"
      operationId: "deleteCity"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of a city"
        required: true
        type: "string"
      responses:
        200:
          description: "OK"
          schema: 
            type: "object"
        400:
          description: "Error"
          schema: 
            type: "object"
...
definitions:
...
  Product_response:
    type: "object"
    properties:
      name:
        type: "string"
      brand:
        type: "string"
      price:
        type: "number"
      options:
        type: "array"
        items:
          type: "object"
          properties: {}
      reviews:
        type: "array"
        items:
          type: "object"
          properties: {}
      id:
        type: "string"
      lastModifiedDate:
        type: "string"
    example:
      reviews:
      - "{}"
      - "{}"
      lastModifiedDate: "lastModifiedDate"
      price: 0.77
      name: "name"
      options:
      - "{}"
      - "{}"
      id: "id"
      brand: "brand"
  Location:
    type: "object"
    properties:
      lat:
        type: "number"
      long:
        type: "number"
    example:
      lat: 0.88888
      long: 5.5533
  City_body:
    type: "object"
    required:
    - "name"
    properties:
      name:
        type: "string"
      country:
        type: "string"
      capital:
        type: "boolean"
        default: false
      location:
        $ref: "#/definitions/Location"
    example:
      country: "country"
      capital: false
      name: "name"
      location:
        lat: 0.99999
        long: 4.44444
  City_response:
    type: "object"
    required:
    - "name"
    properties:
      name:
        type: "string"
      country:
        type: "string"
      capital:
        type: "boolean"
        default: false
      location:
        $ref: "#/definitions/Location"
      id:
        type: "string"
      lastModifiedDate:
        type: "string"
    example:
      country: "country"
      capital: false
      lastModifiedDate: "lastModifiedDate"
      name: "name"
      location:
        lat: 0.99999
        long: 5.55555
      id: "id"
...
host: localhost:8081
basePath: /api/mongo
schemes:
 - http